[BACKPORT 2024.1][#23809] CDCSDK: Filter out records corresponding to index tables in colocated DBs
Summary:
####Backport Description
No merge conflicts were encountered.

####Original Description
If an index is created on a colocated table, the index table also resides on the same colocated tablet. If GetChanges is called on such a tablet, then we may encounter intents corresponding to the index table. Currently for the gRPC model of consumption cdc_service sends such records for index tables and the connector filters them. However for logical replication, the GetChanges call fails when it encounters the intents corresponding to index tables. This is because for logical replication, we try to find the replica identity corresponding to the index table in stream metadata. However since index tables are not part of CDC streams, this operation fails causing GetChanges to return with non-ok status.

In this diff we introduce filtering mechanism by which we filter out the intents corresponding to the index and other non-eligible tables. A record from colocated tablet is shipped from cdc_service if it belongs to a tablet in qualified tables list and does not belong to a table in unqualified tables list.
Jira: DB-12710

Original commit: 361a99af69db667e2ebca977ff9aaa49fc01114c / D37929

Test Plan: ./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestColocationWithIndexes

Reviewers: skumar, siddharth.shah, asrinivasan

Reviewed By: skumar

Subscribers: ycdcxcluster, svc_phabricator, steve.varnau

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37987