[#23881] docdb: Update the hint to advisory_locks
Summary:
With #19974, the requests for advisory locks fail with an error message.
This diff updates the hint in the "advisory locks are not yet implemented"
error message to indicate that there is a workaround in case the app
doesn't need the strict functionality and the GH issue #3642 has the details
of the workaround.
Jira: DB-12785

Test Plan: ./yb_build.sh --cxx-test pgwrapper_pg_row_lock-test --gtest_filter PgRowLockTest.AdvisoryLocksNotSupported

Reviewers: rthallam, smishra, pjain

Reviewed By: rthallam, smishra

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37988