[#23766] docdb: Get colocated schema_version correctly in ChangeMetadataOperation
Summary:
The code was previously always getting the parent table schema version in the apply of ChangeMetadataOperation. This caused CREATE INDEX on a newly created table to fail because the alter on the primary table (to add the index) was ignored because `ChangeMetadataOperation::Apply()` thinks the schema version is already 1 (even though it is 0; 1 is the schema version of the parent table).

This diff changes `ChangeMetadataOperation::Apply()` to get the colocated table's schema version correctly, and refactors the functions to get the schema version so the choice of whether to get the primary table schema version or a colocated table's schema version is more explicit.

The only functional change is in the `ChangeMetadataOperation` and the modified test.
Jira: DB-12667

Test Plan: `./yb_build.sh release --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter *CreateTableAfterClone/1`

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37994