[#23913] build: fix ysql conn mgr ASAN issue
Summary:
Commit aa41478c1a6991b679eda44e0f763eba30ffa6aa / D37829 broke some
ysqlconnmgr ASAN tests, throwing heap-buffer-overflow.  The issue is
with memcpy of 64 bytes even though the source may not have that many
bytes.  This was overlooked because the similar-looking statement before
that was a proper 64 bytes to 64 bytes.  Fix it properly.
Jira: DB-12817

Test Plan:
I was not able to locally reproduce the ASAN failure before this change
on the same alma8-asan-clang17 build setup.  So use jenkins to determine
that it has been fixed.  Also make sure release gcc11 still compiles.

Close: #23913
Jenkins: all tests

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: svc_phabricator, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37998