[#3946] YSQL: Add support for ALTER TABLE ... VALIDATE CONSTRAINT
Summary:
`ALTER TABLE ... VALIDATE CONSTRAINT` is used to validate foreign-key and check constraints that were previously created using the `NOT VALID` syntax.

The command performs the required constraint checking and marks the constraint as valid if successful, or returns an error if there is a violation. It will increment the schema version of the table and any of its children, without making any DocDB schema changes. The increment is done (similar to `ADD CONSTRAINT`, `SET NOT NULL` operations) to abort any concurrent writes with a schema version mismatch error. Note: some writes (that may cause constraint violations) may still slip through, but this is a well known issue that applies to other `ALTER`s as well and it will be fixed by table locking.

Code changes:
- Enable the command in the parser
- YBCPrepareAlterTableCmd: prepare YB alter handles to perform a schema version increment for `AT_ValidateConstraint` and `AT_ValidateConstraintRecurse` commands.
Jira: DB-1922

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestAlterTableWithConcurrentTxn'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressTable'

Reviewers: myang

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D37907