[BACKPORT 2024.1][#23661] YSQL: Change the default wal_level to logical
Summary:
Original commit: 22b6a4730da37612836f17fc29d54790c78db952 / D37930
wal_level is a GUC parameter that isn't applicable in YSQL since we don't use the PG WAL. There are a few logical replication clients which
validate the value of wal_level and ensure that it is set to 'logical' before proceeding with replication.

Those clients can fail when the value of wal_level isn't logical even though it isn't really applicable in YSQL. So for user experience, we are
setting the default value as 'logical' so that the number of user steps is reduced.
Jira: DB-12572

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testDefaultWalLevel'

Reviewers: asrinivasan, skumar, sumukh.phalgaonkar

Reviewed By: skumar, sumukh.phalgaonkar

Subscribers: ycdcxcluster, yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D38013