[DEVOPS-3234] yb_release: Change clean step error to warning
Summary:
In some environments, the "sbt clean" step resolves dependency and successfully downloads the
scala-sbt module, but still returns error and causes the entire build to fail.
This changes an error on the clean step to a warning.

Should not affect official builds that always happen in a clean source tree. A real fatal error
would be detected a bit later when checking for build results.

Test Plan: Jenkins: compile only, build type: release

Reviewers: jmak, kkannan

Reviewed By: jmak

Subscribers: devops, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38023