[PLAT-15261][Master]Backups created through the Scheduled Backup feature are incorrectly tagged under the login user
Summary:
Backups created through the Scheduled Backup feature are incorrectly tagged under the login user (e.g., demo@yugabyte.com) instead of the scheduled backup name. This is because we first check for the Context user instead of the schedule name passed as a param while creating the `CustomerTask` entry.
Switched the order of preference in this diff

Test Plan:
- Create a schedule in a universe
- Verified user in the task page is the schedule name
{F285820}
- Logged out and waited for the backup task to still have schedule name

Reviewers: vkumar

Reviewed By: vkumar

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38029