[PLAT-15300] Update task progress poller logic
Summary:
**Context**
It is possible for the task progress to be at 100%, while not being in failure state or success state.
As a result, the UI will sometimes prematurely send a task completed toast notification even before the task is
completed.

**Change**
The onTaskCompletion function will now only run if the task progress is at 100% and the status is `Success',

Test Plan:
- Run several different tasks that would lock and freeze a universe for update.
  This will create the subtask `createFreezeUniverseTask` which will report a percent completed field of 100% even
  before the actual intended task is completed.
- Verify that the UI toast notification is only fired when the the main task is completed.

Reviewers: rmadhavan, hzare, nsingh

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38030