[#23314] YSQL: Prevent moving a single colocated table to a different tablespace
Summary:
Previously, the `ALTER TABLE <table> SET TABLESPACE <tablespace>` command allowed changing the tablespace of a single colocated table, which should not be possible.
- Colocated tables cannot be moved individually since de-colocation is not supported.
- This fix ensures that attempting to move a single colocated table will now result in an error.
- Instead, to move all tables within a tablespace, use:
```ALTER TABLE ALL IN TABLESPACE <tsp> SET TABLESPACE <new tsp> CASCADE```
- The fix enforces this by setting `yb_cascade` in `AlterTableCmd` to true for the movement of all colocated tables. If `yb_cascade` is not true and a single colocated table is being moved, an error is thrown.

JIRA: DB-12239

Test Plan:
== Automated Testing ==
- ./yb_build.sh --java-test org.yb.pgsql.TestPgRegressColocatedTablesWithTablespaces#testPgRegressColocatedTablesWithTablespaces

In the above test added a scenario, where we try to move a single colocated table in a tablespace, and ensures that it throws an error.

Reviewers: skumar, yguan, aagrawal

Reviewed By: yguan

Subscribers: jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D37055