[#23897] xClusterDDLRepl: Support create table with partition by primary key
Summary:
When creating a table with a partition by the primary key, the primary index created is classified
as RELKIND_PARTITIONED_INDEX instead of RELKIND_INDEX.

Adding support in the ddlrepl extension to handle this type of index as well - since this index is
the same table as the parent table, we don't need to search for / replicate this index.

Fixes #23897.
Jira: DB-12801

Test Plan:
```
ybd --java-test "org.yb.pgsql.TestPgRegressYbExtensionsYbXclusterDdlReplication"
```

Reviewers: xCluster, yyan

Reviewed By: yyan

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38033