[pg15-upgrade] feat: upgrade user databases
Summary: Fix the catalog manager's per-db catalog version logic to match the logic used when per-db catalog version mode is off: use the PG11's catalog versions, which should be frozen, during the upgrade. Without this fix, a new failure mode arises with user-created databases, which is that any pg15 postgres looking for the catalog version map is unable to find a version for user-created databases calling GetSharedCatalogVersion, and the upgrade fails. In the upgrade this happens specifically on the master node which is acting as tserver, with postgres in binary upgrade mode.

Test Plan:
Jenkins

On MacOS 14 arm64:
  ./yb_build.sh release --sj
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh
  ./yb_build.sh fastdebug --sj
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh

Reviewers: hsunder, jhe, myang

Reviewed By: hsunder

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38051