[#20908] YSQL: Introduce interface to optimize index non-key column updates
Summary:
**Background**
Postgres updates tuples by deleting the existing copy of the tuple, and reinserting it with an updated value.
This applies to both relations (tables) as well as indexes. Thus, Postgres' generalized access method interface
includes definitions for inserts and deletes but not updates.

DocDB supports the notion of in-place updates (PGSQL_UPDATE) when the key columns of the tuple remain unmodified.
This mechanism is leveraged while updating tuples of main tables when the primary key of the table remains unmodified.
However, no such mechanism exists for updating index tuples.

**Problem**
Pggate has a buffering mechanism for write requests which allows multiple write requests to be flushed at once, thus
reducing the number of round trips between Postgres and DocDB. To preserve causality/ordering, this buffering mechanism
does not allow multiple writes to the same tuple (identified by the tuple's CTID) to be enqueued together in the buffer.
The previous write to the tuple is flushed before enqueueing the next write.
This poses a problem when updating the non-key columns of indexes. For instance, when the included columns in a covering
index need to be updated, the update is modeled as DELETE + INSERT requests. Since there is no change to the key columns of
the index, this causes the update to require two flushes.

**Solution**
This revision introduces an index update interface in postgres for facilitating in-place updates of index tuples in DocDB.
This replaces the need to perform a DELETE + INSERT on an index update when no key columns are modified.
This is useful in two scenarios:
 - Updating the INCLUDE columns in both unique and non-unique indexes
 - Updating the base table CTID in unique indexes

Prior to updating the index tuple, the execution can check if the index access method supports in-place updates via the
a newly introduced indexAM field `ybamcanupdatetupleinplace`. Currently, in-place updates are only supported for LSM indexes.

Note - The base table CTID is a part of the index key for non-unique indexes.

**Example**
Consider a table containing two covering indexes, one unique and one non-unique.
The example below displays
```
yugabyte=# \d demo
                Table "public.demo"
 Column |  Type   | Collation | Nullable | Default
--------+---------+-----------+----------+---------
 k      | integer |           | not null |
 v1     | integer |           |          |
 v2     | integer |           |          |
 v3     | integer |           |          |
 v4     | integer |           |          |
Indexes:
    "demo_pkey" PRIMARY KEY, lsm (k HASH)
    "demo_v2_v4_idx" UNIQUE, lsm (v2 HASH) INCLUDE (v4)
    "demo_v1_v3_idx" lsm (v1 HASH) INCLUDE (v3)

yugabyte=# SET yb_explain_hide_non_deterministic_fields TO true;

-- Query 1: Update of included columns in non-unique index
yugabyte=# EXPLAIN (ANALYZE, DIST) UPDATE demo SET v3 = v3 + 1 WHERE k = 1;
                                             QUERY PLAN
-----------------------------------------------------------------------------------------------------
 Update on demo  (cost=0.00..4.12 rows=1 width=96) (actual rows=0 loops=1)
   ->  Index Scan using demo_pkey on demo  (cost=0.00..4.12 rows=1 width=96) (actual rows=1 loops=1)
         Index Cond: (k = 1)
         Storage Table Read Requests: 1
         Storage Table Rows Scanned: 1
         Storage Table Write Requests: 1
         Storage Index Write Requests: 1       // would have been 2 previously
 Storage Read Requests: 1
 Storage Rows Scanned: 1
 Storage Write Requests: 2
 Storage Flush Requests: 1        // would have been 2 previously
(11 rows)

-- Query 2: Update of primary key + included columns in unique index
yugabyte=# EXPLAIN (ANALYZE, DIST) UPDATE demo SET k = k + 1, v4 = v3 + 1 WHERE k = 10;
                                             QUERY PLAN
-----------------------------------------------------------------------------------------------------
 Update on demo  (cost=0.00..4.12 rows=1 width=96) (actual rows=0 loops=1)
   ->  Index Scan using demo_pkey on demo  (cost=0.00..4.12 rows=1 width=96) (actual rows=1 loops=1)
         Index Cond: (k = 10)
         Storage Table Read Requests: 1
         Storage Table Rows Scanned: 1
         Storage Table Write Requests: 2
         Storage Index Write Requests: 3        // would have been 4 previously
 Storage Read Requests: 1
 Storage Rows Scanned: 1
 Storage Write Requests: 5
 Storage Flush Requests: 1        // would have been 2 previously
(11 rows)
```
Previously, both of the above cases would have required two flushes as the key columns of the respective indexes remain unchanged across the DELETE and INSERT requests.

**Future Work**
 - The computation of whether key columns of an index are modified, can be performed at planning time.
    This can be done once D34040 lands.
 - Refactor ApplyUpdate in pgsql_operation.cc to delineate table update (main table, index) and sequence update operations.
Jira: DB-9891

Test Plan:
Run the following tests:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressUpdateOptimized#schedule'
```

Reviewers: amartsinchyk, tnayak, jason

Reviewed By: amartsinchyk, tnayak

Subscribers: ybase, smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36588