[PLAT-15320][YBA CLI]K8s Universes throw `interface {} is string, not []float64` error
Summary: After the introduction of template files for configuration, the float slices conversion started throwing an error. This is because viper does not contain a function to read float slices, and we ha to rely on viper.Get() to fetch the values. Despite the return type being interface, the actual type of the Get() return is string, which is what was causing the current issue. Introduced a function that would explicitly convert the string read by `viper.Get()` to a float slice in this diff.

Test Plan: Tested out k8s universe creation using YBA CLI

Reviewers: sneelakantan

Reviewed By: sneelakantan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38065