Merged PR 1458853: [Cleanup] Merge Legacy Cleanup PR
### Does this PR have any customer impact?

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

### Are you introducing any new config? If yes, do you have tests with and without them being set?

### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description
Before 1.20 release join condition we were not pushing down to index so in join function itself we use to validate `on` field values and compare target with source. but In 1.20 we have introduced index index push down for join condition where we are using support function for `bson_dollar_merge_join` so we validate on fields values in source projection itself using `bson_dollar_extract_merge_filter`.

So now as part of this PR we are at 1.20+ so removed older code bits which we are not going to use.

this was the PR where we have introduced index pushdown
https://msdata.visualstudio.com/CosmosDB/_git/pgmongo/pullrequest/1425893