[#23940] YSQL: Replace copyright string from YugaByteDB to YugabyteDB
Summary:
When running arc lint on a file that has

```
// Copyright (c) YugaByteDB, Inc.
```

the linter fails with an error

```
arc lint
 Exception
Undefined offset: -1
(Run with `--trace` for a full exception trace.)
```

The error message is not informative since this is an array index out of bounds exception within the linter. This makes it difficult for devs to root cause the issue.

This revision replaces the copyright string from the old version `YugaByteDB` to the correct version `YugabyteDB`.
Jira: DB-12838

Test Plan: Jenkins

Reviewers: amartsinchyk

Reviewed By: amartsinchyk

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38073