[#23956] YSQL: Fix org.yb.pgsql.TestYsqlMetrics#testMetricRows
Summary:
D37636 adds table-level metrics to the `/statements` endpoint.
This means that many metrics have the same name but a different
`table_name` field.

The parser in `Metrics.java` expects there to be only one metric
with a given name, so if it sees the same name more than once, it
overwrites the previous value for that metric.

This revision modifies the metrics parser to skip over table-level metrics.
Jira: DB-12855

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestYsqlMetrics#testMetricRows'
```

Reviewers: myang

Reviewed By: myang

Differential Revision: https://phorge.dev.yugabyte.com/D38103