[PLAT-15288] Use set_dbs endpoint when editing table selection for db scoped DR configs
Summary:
For DB scoped DR configs, we have a different endpoint for the edit table flow.
Instead of set_tables, the client should send a request to the `set_dbs` endpoint instead.

This diff also modifies the form field for the set_dbs request and edit xCluster config form by changing the
`databases` to `dbs`. This change is made to stay consistent with the `dbs` field used during db scoped DR creation.

Test Plan:
Verify that user is able to add databases to and remove databases from a db scoped DR config.
Verify that the YBA UI still uses set_tables endpoint when the config is not db scoped.

Reviewers: rmadhavan, cwang, hzare, vbansal

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38026