[#24005] DocDB: Table Locks - check for conflicts at master.
Summary:
This diff adds a TSLocalLockManager to the ObjectLockInfoManager at master.
This allows master to check and detect DDL lock requests that may end up blocking on other DDL requests.

Any request which is accepted by Master's local lock manager, can be immediately persisted, even before forwarding the lock requests to all the TServers. This simplifies the bootstrap process of new TServers, as they do not have to worry about any in-progress requests. -- any inprogress request would have already persisted it's state at the master, and been included in the bootstrap.

One con of this approach is that -- until we implement wait-queues/async queues -- waiting for the lock may involve blocking an RPC thread. This should be few in number as we do not expect to have too many DDL requests.
Jira: DB-12893

Test Plan: ybd --cxx-test object_lock-test

Reviewers: bkolagani, zdrudi

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38167