[#23943]: YSQL: Fix Bitmap Scan crash in fastdebug GCC11
Summary:
Intermittently, in fastdebug compiled with gcc11 (no repro on other build types, or on fastdebug with clang), bitmap scan queries with a Bitmap And crash.

The crash is caused by a segmentation fault while removing an item from the bitmap set:
```lang=c++
  // for each elem in a, if it's not also in b, delete a's copy
  auto iterb = b->begin();
  for (auto itera = a->begin(); itera != a->end();) {
    if ((iterb = b->find(*itera)) == b->end()) {
      FreeSlice(*itera);
      itera = a->erase(itera); // <----- crashes here inside the erase call.
    } else {
      ++itera;
    }
  }
```

Moving the `FreeSlice` line to after removing the item from the list allows it to complete successfully, with the same behaviour as before. My guess is that gcc11 tries to invoke the destructor of the slice when removing it from the set, and since the slice is freed but not set to NULL, it points to some garbage value that occasionally results in a crash.

Test Plan:
```lang=sj
./yb_build.sh fastdebug --java-test 'org.yb.pgsql.TestPgRegressYbBitmapScans' --gcc11
```

Or, manually:
1. Build:
```
./yb_build.sh fastdebug --gcc11
```

2. Restart cluster

3. Run
```lang=sql
/*+ Set(yb_enable_bitmapscan true) Set(yb_enable_base_scans_cost_model true) BitmapScan(t) */
EXPLAIN (ANALYZE, SUMMARY OFF, COSTS OFF) SELECT * FROM test_and t WHERE a < 5 AND b < 5;

\watch .1
```

Before this change, it would crash usually on the first or second invocation. With this change, it runs successfully for at least a few minutes.

Reviewers: amartsinchyk

Reviewed By: amartsinchyk

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38080