[#23975] YSQL, ASH: Fix TestYbAsh#testPgAuxInfo
Summary:
Earlier, this test used to run a bunch of queries with
high sampling rate to make sure that the circular buffer
would wrap around. This started failing after tests in
tsan were enabled. Now the size of the circular buffer
is reduced to 1 KiB from the default of 16 KiB for the
test and the sampling interval is updated to 50 ms.
Jira: DB-12876

Test Plan:
Jenkins: test regex: testPgAuxInfo

The test passed 240/240 times with tsan on dev server

Reviewers: jason

Reviewed By: jason

Differential Revision: https://phorge.dev.yugabyte.com/D38122