[#23741] docdb: Add parameter to run PITR tests with clone instead
Summary:
This diff adds a `RestoreType` parameter to some PITR tests in `yb-admin-snapshot-schedule-test` to test them with DB cloning. If `RestoreType` is set to `RestoreType::kClone`, `RestoreSnapshotSchedule` will clone to a separate database as of the restore time, instead of restoring over the source database. We now also refresh the connection with `ConnectToRestoredDb()`, which will connect to the appropriate database (the source db if using PITR, or the target db if using clone).

Fixes #23741

Jira: DB-12646

Test Plan: `./yb_build.sh build release --cxx-test yb-admin-snapshot-schedule-test`

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37723