[#24123] YSQL: Make a new ANN Access Method that redirects to the base table
Summary:
This change introduces infrastructure for creating indexes that are "covered" by the main table much like a primary key index. This is achieved by adding a boolean variable to the `IndexAmRoutine` struct by the name of `yb_amiscoveredbymaintable` to denote whether or not the index is inlined with the underlying table. YB code that has traditionally checked `indisprimary` to avoid doing operations to the main table is replaced by checks to a utility function `YBIsCoveredByMainTable(Relation)` that checks `indisprimary` along with `yb_amiscoveredbymaintable`.

This work is done in preparation for vector indexes that are colocated with an underlying table. This change introduces the index access method `yblocalann` as an example of such an index type. This index AM can only be created on colocated tables for now as there isn't infrastructure to read ordered results from a multi-tablet vector index yet.

**Upgrade/Rollback safety:**
This change moves where `PgVectorReadOptionsPB` is defined and adds a protobuf field to that `vector_column_id` to signify what column is being used as the vector key. This shouldn't affect any user of the previous format. This is also for a use case that is not GA or EA.
Jira: DB-13014

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsPgvector'

Reviewers: fizaa, tfoucher

Reviewed By: fizaa

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38176