[PLAT-15378][localProvider][dr] Deflake testDrConfigSetup local provider test
Summary:
This diff adds retry logic around the check for number of rows in the target universe in an xCluster setup. Because an xCluster replication is asynchronous, it retries for 1 minute to check if the inserted row on the source universe shows up on the target universe.

It also disables `com.yugabyte.yw.commissioner.tasks.local.DRDbScopedLocalTest#testDrDbScopedUpdate` local provider test.

Test Plan:
- Made sure the retry logic works properly when the delay is 10ms.
- Made sure testDrConfigSetup passes

Reviewers: cwang, vbansal, sanketh

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38190