[PLAT-14460] Implementing restore from s3 backup
Summary: Diff implements support for restoring from s3 backup. Based on the storage config and backup directory provided we query for the most recent backup in that folder and restore it to YBA. Afterwards, we perform a shutdown so any migrations and postgres connections are reset.

Test Plan: Restore from s3 backup, ensure success

Reviewers: vkumar, sanketh, dshubin, vbansal

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38245