[PLAT-15338] Improve formatting for very long universe names in DR participant cards
Summary:
This diff changes the width for DR participant cards to min-width which allows the card to grow horizontally to
accomodate for very long univese names.

Test Plan:
Verify the DR participant card width grows as expected when given longer universe names.
Verify the overall UI layout is as expected.
{F287870}

Reviewers: rmadhavan, cwang, hzare

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38247