[BACKPORT 2024.2][#23988] YSQL: Skip a table for schema comparison if its schema does not change
Summary:
There is a flaw with the current DDL atomicity work flow based upon schema
comparison when the DDL transaction involves multiple tables. Only the first
table's schema is used for comparison on the grounds that we only need to
compare one table's schema against PG catalog to determine the DDL transaction's
committed/aborted status. While this assumption is generally true, it fails when
the first table's schema does not change and only involves a schema version
increment.

For example, a DDL like `alter table mytable add constraint x_unique unique(x)`
does not change the table `mytable`'s DocDB schema, but it creates a new index
`x_unique` in DocDB. We cannot use `mytable` for schema comparison to decide
whether the transaction has committed or aborted because its schema does not
change. But we can use `x_unique` for that purpose.

To fix the bug, I made a change in `TriggerDdlVerificationIfNeeded` to skip
a table for schema comparison when there are multiple tables involved in
the DDL transaction and the table is already selected once and its schema
does not change before and after the DDL. In this way, we will skip `mytable`
after selecting it once and pick `x_unique` for schema comparison, and will
be able to decide whether the DDL transaction has committed or aborted.

If all the tables schema do not change (e.g., alter table set a check constraint),
we will not keep picking the same table for schema comparison because that will
result in infinite loop. In this case it does not matter whether the DDL transaction
has committed or aborted. We only need to know that the DDL transaction has
terminated, which we do and we can treat the DDL transaction as either committed
or aborted.
Jira: DB-12883

Original commit: 4a07eb8c1b50ad94d1638092cd302434610fddc3 / D38250

Test Plan:
./yb_build.sh debug --cxx-test pg_ddl_atomicity-test --gtest_filter PgDdlAtomicityTest.TestAlterTableAddUniqueConstraint
./yb_build.sh debug --cxx-test pg_ddl_atomicity-test --gtest_filter PgDdlAtomicityTest.TestAlterTableAddCheckConstraint

Reviewers: fizaa

Reviewed By: fizaa

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38619