[PLAT-15396] Fix up update db scoped replication local provider test
Summary:
Re-enable `testDrDbScopedUpdate` local provider test, as it is an issue with db. When db is dropped from db scoped replication, it takes time for masters to propagate this information to the tservers.

In this local provider test, right after drop db from replication, we immediately add a row to a table in the db that was dropped. We add a sleep timeout to make sure that the tservers are notified before inserting the rows.

Test Plan: Local provider test passing (adding extra check to make sure row is not replicated to target) + checking verbose logging of local provider test.

Reviewers: hzare, vbansal, sanketh

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38251